feat: Input View 기능 이동
